+++ client/app/lib/Views/giver/register/category_sheet.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import 'package:app/Views/components/color_cell.dart';
+import '../../common/cateogory_cell.dart';
 
-class ColorSheet extends HookConsumerWidget {
-  const ColorSheet({
+class CategorySheet extends HookConsumerWidget {
+  const CategorySheet({
     super.key,
   });
 
@@ -31,7 +31,7 @@ class ColorSheet extends HookConsumerWidget {
                 icon: const Icon(Icons.arrow_back_ios),
               ),
               const Text(
-                'カラー',
+                'カテゴリー',
                 style: TextStyle(
                   fontWeight: FontWeight.bold,
                 ),
@@ -49,30 +49,30 @@ class ColorSheet extends HookConsumerWidget {
                 Row(
                   mainAxisAlignment: MainAxisAlignment.spaceBetween,
                   children: [
-                    ColorCell(colorIndex: 0, isSheet: true),
-                    ColorCell(colorIndex: 1, isSheet: true),
-                    ColorCell(colorIndex: 2, isSheet: true),
-                    ColorCell(colorIndex: 3, isSheet: true),
+                    CategoryCell(categoryIndex: 0, isSheet: true),
+                    CategoryCell(categoryIndex: 1, isSheet: true),
+                    CategoryCell(categoryIndex: 2, isSheet: true),
+                    CategoryCell(categoryIndex: 3, isSheet: true),
                   ],
                 ),
                 SizedBox(height: 12),
                 Row(
                   mainAxisAlignment: MainAxisAlignment.spaceBetween,
                   children: [
-                    ColorCell(colorIndex: 4, isSheet: true),
-                    ColorCell(colorIndex: 5, isSheet: true),
-                    ColorCell(colorIndex: 6, isSheet: true),
-                    ColorCell(colorIndex: 7, isSheet: true),
+                    CategoryCell(categoryIndex: 4, isSheet: true),
+                    CategoryCell(categoryIndex: 5, isSheet: true),
+                    CategoryCell(categoryIndex: 6, isSheet: true),
+                    CategoryCell(categoryIndex: 7, isSheet: true),
                   ],
                 ),
                 SizedBox(height: 12),
                 Row(
                   mainAxisAlignment: MainAxisAlignment.spaceBetween,
                   children: [
-                    ColorCell(colorIndex: 8, isSheet: true),
-                    ColorCell(colorIndex: 9, isSheet: true),
-                    ColorCell(colorIndex: 10, isSheet: true),
-                    ColorCell(colorIndex: 11, isSheet: true),
+                    CategoryCell(categoryIndex: 8, isSheet: true),
+                    CategoryCell(categoryIndex: 9, isSheet: true),
+                    CategoryCell(categoryIndex: 10, isSheet: true),
+                    CategoryCell(categoryIndex: 11, isSheet: true),
                   ],
                 ),
                 SizedBox(height: 24),
