+++ client/app/lib/Views/giver/register/register_product_sheet.dart
@@ -1,18 +1,17 @@
 import 'dart:io';
-
 import 'package:camera/camera.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import 'package:app/constants.dart';
-import 'package:app/Usecases/provider.dart';
-import 'package:app/Views/components/color_sheet.dart';
-import 'package:app/Views/components/category_sheet.dart';
-import 'package:app/Views/components/condition_sheet.dart';
-import 'package:app/Views/components/register_trade_sheet.dart';
-import 'package:app/Views/components/register_picture_sheet.dart';
+import '../../../Domain/constants.dart';
+import '../../../Usecases/provider.dart';
+import 'color_sheet.dart';
+import 'category_sheet.dart';
+import 'condition_sheet.dart';
+import 'register_trade_sheet.dart';
+import 'register_picture_sheet.dart';
 
 class RegisterProductSheet extends HookConsumerWidget {
   final ValueNotifier<CameraController?> cameraController;
