+++ client/app/lib/Views/search/keyword_search_view.dart
@@ -3,8 +3,8 @@ import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
-import 'package:app/Views/search_result_view.dart';
-import 'package:app/Views/components/cateogory_cell.dart';
+import '../common/cateogory_cell.dart';
+import 'search_result_view.dart';
 
 class KeywordSearchView extends HookConsumerWidget {
   const KeywordSearchView({super.key});
