+++ client/app/lib/Views/search/picture_search_view.dart
@@ -3,9 +3,9 @@ import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import 'package:app/constants.dart';
-import 'package:app/Usecases/provider.dart';
-import 'package:app/Views/components/cateogory_cell.dart';
+import '../../Domain/constants.dart';
+import '../../Usecases/provider.dart';
+import './../common/cateogory_cell.dart';
 
 class PictureSearchView extends HookConsumerWidget {
   final ValueNotifier<CameraController?> cameraController;
